DashboardBackend/requirements.txt
kafka-python
flask
flask_socketio
psycopg2-binary
